telegram_bot_service/requirements.txt
telethon
sqlalchemy==2.0.20
psycopg2-binary==2.9.9
python-dotenv==1.0.1